Possible modifier rework example
Essentially from what I gathered the modifiers/interfaces such as healable and flammable require a new script for each node they'd want to be placed on. I began messing around with signals to see if we could improve code modularity and improve separation of concerns. It would honestly be a pain to rework everything to use this but could be useful and reduce the need for a new script every time we have a new obj that needs healing such as the ballista. Currently got it working for the bridge and player but as this was just an example of we could I stopped there